move header to shared
