+++ websites/mswjs.io/package.json
@@ -14,32 +14,12 @@
     "url": "https://github.com/mswjs/mswjs.io/tree/main/websites/mswjs.io"
   },
   "dependencies": {
-    "@astrojs/markdown-remark": "^3.2.0",
-    "@astrojs/mdx": "^1.1.2",
-    "@astrojs/react": "^3.0.3",
-    "@astrojs/rss": "^3.0.0",
-    "@astrojs/sitemap": "^3.0.2",
-    "@astrojs/tailwind": "^5.0.2",
-    "@docsearch/css": "^3.5.2",
-    "@docsearch/react": "^3.5.2",
-    "@heroicons/react": "^2.0.18",
-    "@tailwindcss/typography": "^0.5.10",
-    "@types/body-scroll-lock": "^3.1.0",
-    "@types/node": "^20.8.0",
-    "@types/react": "^18.0.21",
-    "@types/react-dom": "^18.0.6",
-    "astro": "^3.3.2",
-    "body-scroll-lock": "4.0.0-beta.0",
     "dotenv": "^16.3.1",
     "glob": "^10.3.10",
     "gray-matter": "^4.0.3",
     "minimatch": "^9.0.3",
-    "outvariant": "^1.4.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "rehype-autolink-headings": "^7.0.0",
-    "rehype-external-links": "^3.0.0",
-    "tailwindcss": "^3.0.24"
+    "mswjs.io-shared": "workspace:*",
+    "outvariant": "^1.4.0"
   },
   "devDependencies": {
     "@astrojs/markdoc": "^0.5.0",
