+++ websites/mswjs.io/src/components/PageBanner.astro
@@ -1,5 +1,5 @@
 ---
-import { Container } from "./react/container";
+import { Container } from "mswjs.io-shared/components/react/container";
 ---
 
 <aside class="py-1.5 border-b border-neutral-700 bg-neutral-950 text-neutral-200 text-sm text-center font-medium">
