+++ websites/mswjs.io/src/consts.ts
@@ -1,5 +1,4 @@
 export const SITE_TITLE = 'Mock Service Worker'
 export const SITE_DESCRIPTION = 'API mocking library for browser and Node.js'
-export const DISCORD_URL = 'https://kettanaito.com/discord'
 export const EGGHEAD_COURSE_URL =
   'https://egghead.io/courses/mock-rest-and-graphql-apis-with-mock-service-worker-8d471ece'
