+++ websites/mswjs.io/src/content/docs/getting-started.mdx
@@ -168,8 +168,8 @@ import { BugAntIcon } from '@heroicons/react/24/outline'
 
 We are always happy to help you with any questions you might have. There are multiple ways you can reach out to us or other community members to get help:
 
-import { DISCORD_URL } from '../../consts'
-import { DiscordIcon } from '../../components/react/icons/discord'
+import { DISCORD_URL } from 'mswjs.io-shared/consts'
+import { DiscordIcon } from 'mswjs.io-shared/components/react/icons/discord'
 
 - <a href={DISCORD_URL} target="_blank" rel="noopener noreferrer">
     <DiscordIcon className="inline mr-1.5 w-5 align-text-top" />
