+++ websites/mswjs.io/src/pages/404.astro
@@ -1,10 +1,10 @@
 ---
 import BaseHead from "../components/BaseHead.astro";
-import Header from "../components/Header.astro";
-import { Container } from "../components/react/container";
-import { Footer } from "../components/react/footer";
-import { Grid } from "../components/react/grid";
-import { LinkExternal } from "../components/react/linkExternal";
+import Header from "mswjs.io-shared/components/Header.astro";
+import { Container } from "mswjs.io-shared/components/react/container";
+import { Footer } from "mswjs.io-shared/components/react/footer";
+import { Grid } from "mswjs.io-shared/components/react/grid";
+import { LinkExternal } from "mswjs.io-shared/components/react/linkExternal";
 import { SITE_DESCRIPTION, SITE_TITLE } from "../consts";
 
 const { pathname } = Astro.url
