+++ websites/shared/components/react/icons/youtube.tsx
