+++ websites/shared/consts.ts
@@ -1,0 +1,1 @@
+export const DISCORD_URL = 'https://kettanaito.com/discord'
