+++ websites/shared/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "mswjs.io-shared",
+  "version": "0.0.0"
+}
